`pop.var` argument
added pop.var argument that is necessary for computation of rates. I also updated debugged all subsequent code that involves that argument. Added an example using the `pop.var` argument. This commit also adds the new functionality where if `pop.var` and `add.summary.row` or `add.summary.col` are TRUE and "rate" is selected in `metric`, the output table does not produce rates for the summary row and column in the table. More examples were added as unit tests